Add global exception handling

+++ student-api/src/main/java/com/sarthak/student_api/service/StudentService.java
@@ -1,13 +1,13 @@
 package com.sarthak.student_api.service;
 
 
+import com.sarthak.student_api.exception.StudentNotFoundException;
 import com.sarthak.student_api.model.Student;
 import com.sarthak.student_api.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class StudentService {
@@ -18,15 +18,15 @@ public class StudentService {
     public List<Student> getAllStudents(){
         return studentRepository.findAll();
     }
-    public Optional<Student> getStudentById(long id){
-        return studentRepository.findById(id);
+    public Student getStudentById(long id){
+        return studentRepository.findById(id).orElseThrow(() -> new StudentNotFoundException(id));
     }
     public Student addStudent(Student student){
         return studentRepository.save(student);
     }
 
     public Student updateStudent(Long id,Student updatedStudent){
-        Student existing = studentRepository.findById(id).orElseThrow(() -> new RuntimeException("Student not found of Id: "+id));
+        Student existing = studentRepository.findById(id).orElseThrow(() -> new StudentNotFoundException(id));
 
         existing.setName(updatedStudent.getName());
         existing.setEmail(updatedStudent.getEmail());
@@ -36,6 +36,9 @@ public class StudentService {
     }
 
     public String deleteStudent(Long id){
+        if (!studentRepository.existsById(id)){
+            throw new StudentNotFoundException(id);
+        }
         studentRepository.deleteById(id);
         return "Student with id "+id+" deleted successfully";
     }
